Remove BaseGuildEmojiManager
